@hovey found a small but critical fix in the Exodus file writer (within the function reorder_connectivity which converts our connectivity structure into one for the NetCDF writer). So the "Element-to-node connectivity" part of the Writing section for Exodus files is now maybe 30% faster.

Changed our net rate for the blocks benchmark comparisons from 5.5 million voxels/second to just over 6.9 million voxels/second. So we are about 70x faster than SCULPT now in this case. It is about 50/50 for us on Meshing vs. Writing for Exodus files. For other files, Writing is of course the bottleneck for direct meshing.

Before:

<pre><font color="#58C7E0">cargo</font> <font color="#58C7E2">run</font> <font color="#58C7E2">-qr</font> <font color="#58C7E2">-F</font> <font color="#58C7E2">profile</font> <font color="#58C7E2">--</font> <font color="#58C7E2">mesh</font> <font color="#58C7E2">hex</font> <font color="#58C7E2">-i</font> <font color="#58C7E2"><u style="text-decoration-style:solid">benches/block/block_720.npy</u></font> <font color="#58C7E2">-o</font> <font color="#58C7E2">target/block_720.exo</font>
<font color="#EC00FF"><b>    automesh 0.3.3</b></font>
     <font color="#58C7E2"><b>Reading</b></font> benches/block/block_720.npy
        <font color="#54E484"><b>Done</b></font> 154.74406ms [1 materials, 373248000 voxels]
     <font color="#58C7E2"><b>Meshing</b></font> voxels into hexes
           <font color="#E0B401"><b>⤷ Removed voxels</b></font> 9.378739216s
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 7.285224451s
             <font color="#E0B401"><b>Nodal coordinates</b></font> 5.722238686s
             <font color="#E0B401"><b>Renumbered nodes</b></font> 4.117404616s
        <font color="#54E484"><b>Done</b></font> 27.499412654s [1 blocks, 373248000 elements, 374805361 nodes]
     <font color="#58C7E2"><b>Writing</b></font> target/block_720.exo
           <font color="#E0B401"><b>⤷ Element-to-node connectivity</b></font> 19.973423787s
             <font color="#E0B401"><b>Nodal coordinates</b></font> 9.127398218s
        <font color="#54E484"><b>Done</b></font> 31.592302436s
       <font color="#EC00FF"><b>Total</b></font> 59.882295263s</pre>

After:

<pre><font color="#58C7E0">cargo</font> <font color="#58C7E2">run</font> <font color="#58C7E2">-qr</font> <font color="#58C7E2">-F</font> <font color="#58C7E2">profile</font> <font color="#58C7E2">--</font> <font color="#58C7E2">mesh</font> <font color="#58C7E2">hex</font> <font color="#58C7E2">-i</font> <font color="#58C7E2"><u style="text-decoration-style:solid">benches/block/block_720.npy</u></font> <font color="#58C7E2">-o</font> <font color="#58C7E2"><u style="text-decoration-style:solid">target/block_720.exo</u></font>
<font color="#EC00FF"><b>    automesh 0.3.3</b></font>
     <font color="#58C7E2"><b>Reading</b></font> benches/block/block_720.npy
        <font color="#54E484"><b>Done</b></font> 145.875868ms [1 materials, 373248000 voxels]
     <font color="#58C7E2"><b>Meshing</b></font> voxels into hexes
           <font color="#E0B401"><b>⤷ Removed voxels</b></font> 8.780411386s
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 6.739418613s
             <font color="#E0B401"><b>Nodal coordinates</b></font> 5.657648441s
             <font color="#E0B401"><b>Renumbered nodes</b></font> 4.041100239s
        <font color="#54E484"><b>Done</b></font> 26.161402382s [1 blocks, 373248000 elements, 374805361 nodes]
     <font color="#58C7E2"><b>Writing</b></font> target/block_720.exo
           <font color="#E0B401"><b>⤷ Element-to-node connectivity</b></font> 12.143821757s
             <font color="#E0B401"><b>Nodal coordinates</b></font> 7.818196663s
        <font color="#54E484"><b>Done</b></font> 27.579682933s
       <font color="#EC00FF"><b>Total</b></font> 54.480233588s</pre>